improved the code structure to allow for more explicit ordering of how files are closed and opened.
if Vis are being controlled then they wait for the file close action button to change, otherwise they prompt. I am
having issues with hte event structure that is waiting for those changes
registering the previous value because it also registers the changes
when the save data button is turned off. Possibly need to move to
notifiers to gain explicit FIFO capability.


Former-commit-id: 7b1c257eddc916e17dd5af1e9784b9561b01e832