separate UI layout code
the UI layout code is now in a separate python file that is imported,
saving space but also allowing for easy updating of the UI layout. to
update the layout, modify the layout.ui file in qt designer and then run
the Convert_ui2py.bat file, which will create a new layout.py file.