restructuring and change MHDpy importing
changed the file structure to move eventlog and post processing routines
to the MHDpy module. previously there as a snippet of code required in
each python file adding the repository path to sys.path. This was
problematic because the path would have to manually be updated. now the user runes setup.bat which will add
the current directory (repository path) to the site-packages of the
detected python installation. The bat file adds a '.pth' file which
simply adds the listed directories in the .pth file to sys.path anytime
the program is run. Also, the bat file installs requirements.txt. also
updated reademe.md