working setup of __init__ files, good enough
cant figure out how to get it as dynamic as I think I'd like. Having to
be explicit with the importing of modules. Think it would be best to
have dynamic importing of all modules in a folder using relative paths but can't figure that
out. Having to delcare and __all__ variable and use absolute paths...