add a button to reduce the number of datapoints for large files
this uses a 'stride' to reduce the size of the internal data and time
array (used in stats and time calculations) as well as the arrays that
ar pulled from plotting. If the box is not ticked then stride is 1 and
the full arrays are returned. If it is ticked then the stride is set so
that there is 1000 datapoints in the arrays.